server/requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
anyjson==0.3.3
argparse==1.2.1
gevent==1.0
gevent-socketio==0.3
gevent-websocket==0.9
greenlet==0.4.1
itsdangerous==0.23
python-mpd==0.3.0
wsgiref==0.1.2
